Migration and Laravel Localization
Migration: created a database its name is ecomercelaravel. so that, we can migrate and create our tables. such as, users table which enabled us to register and save our users data into that table and same for the other tables that have its own jobs.
Laravel Localization: this e-commerce will be able to be browsed in multiple languages(English and Arabic). so, for us to be able to do that we had to install the mcamara package via composer, then edit the routes middleware at the kernal.php file. then edited our route at web.php file. then at config\laravellocalization.php we uncommented the languages we decided the  e-commerce will support.