Customer Dashboard
make login and register for customer and enabling the idea of multi-authentication. here the user will choose when press login whither he would like to login as an admin or customer. and sure if we don’t to do that and make it all for customer and specify a special different route for admin to log in through.