Category
enable admin to add(create) category with attached photo. after of course making a table for it in DB and he can also edit it and delete it with easy restore using SoftDelete feature.
we create an category index to show the admin all the categories he has and enable him to edit or delete these categories. added the category section at the home page for customers to be able to browser it.
-made a MVC dedicated for Category.
-used the trait AttachFileTrait at CategoryController.
-at Category.php used SoftDelete
- at filesystem.php at disks added new folder its name is upload_attachments.
-created admin.php at routes and added to it all admin’s routes.